http-infoDev/requirements.txt
asn1crypto==1.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==2.3.0
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
deap==1.3.0
future==0.18.0
idna==2.2
keyring==10.1
keyrings.alt==1.3
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.17.2
onboard==1.4.1
pexpect==4.2.1
Pillow==4.0.0
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.22.0
pyparsing==2.4.2
pysmbc==1.0.15.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-telegram-bot==12.1.1
pyxdg==0.25
redis==3.3.10
reportlab==3.3.0
requests==2.12.4
SecretStorage==2.3.1
six==1.12.0
tornado==6.0.3
urllib3==1.19.1
